Updated to use integral position values with JUI::Graphics2D. Fixed many missing return values, fixed bad initialization of notifications, fixed missing virtual keyword on some base class destructors.

+++ dragged_slot_view.cpp
@@ -53,7 +53,7 @@ JUI::IOResult DraggedSlotView::on_mouse_moved( JUI::Mouse* mouse )
 /*
  * Set offset to mouse.
  */
-void DraggedSlotView::set_offset( float x, float y )
+void DraggedSlotView::set_offset( int x, int y )
 {
 	offset_x_ = x;
 	offset_y_ = y;
